requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
psycopg2-binary>=2.9.0
ollama>=0.0.6
sentence-transformers>=2.2.0
numpy>=1.24.0
python-multipart>=0.0.5